for hacktoberfest

+++ arr.cpp
@@ -6,7 +6,7 @@ using namespace std;
 int main() {
 	 int N;
 	cin>>N;
-	long long  int *a = new long long int[N];
+	  int *a = new int[N];
 	for(int i =0; i <N; i++ ){
 		cin>>a[i];
 	}
